🚒 fix crash by removing android tracing subscriber
Doublie initialization of tracing subscribters will panic the rust
binaries and therefore crash the android app. To get rid of this crash
we need to stop initializing twice. Since we already made plans to
replace Android tracing with subscriber with a custom callback
impelemntation to get access, to log-messages on app level removing it
is the obvious fix here.