🐛 fix starting arti on android
The native method for starting Arti declared socks_port and dns_port to
be of type Integer on the java side. However the implementation on the
rust side declared those values to be of type jint. So it falsely used
the pointer values (to those java objects of type Integer) instead of
values.